requirements.txt
ipykernel
langchain
python-dotenv
langchain_community

wikipedia

langchain-text-splitters
pandas
duckdb
python-dotenv
sentence_transformers
langchain_huggingface
langchain_chroma
streamlit
langchain-ollama
langchain-groq
langchain_core
huggingface_hub
validators==0.28.1
unstructured
numpy
scipy
numexpr
langchain-experimental
langgraph
speechrecognition 
pyaudio
pillow
pytesseract
